add status badge to candidate page header
