+++ apps/web/src/components/AssessmentStepTracker/utils.ts
@@ -10,17 +10,19 @@ import {
   PoolCandidate,
   ArmedForcesStatus,
   AssessmentDecision,
-  AssessmentResult,
   Maybe,
-  PoolSkillType,
   AssessmentStep,
-  AssessmentStepType,
-  AssessmentResultType,
 } from "@gc-digital-talent/graphql";
-import { notEmpty, unpackMaybes } from "@gc-digital-talent/helpers";
+import { notEmpty } from "@gc-digital-talent/helpers";
 
 import { NO_DECISION, NullableDecision } from "~/utils/assessmentResults";
 import poolCandidateMessages from "~/messages/poolCandidateMessages";
+import {
+  ResultDecisionCounts,
+  determineCandidateStatusPerStep,
+  determineCurrentStepPerCandidate,
+  getDecisionCountForEachStep,
+} from "~/utils/poolCandidate";
 
 export type CandidateAssessmentResult = {
   poolCandidate: PoolCandidate;
@@ -106,24 +108,6 @@ export const getDecisionInfo = (
   };
 };
 
-export type ResultDecisionCounts = Record<NullableDecision, number>;
-
-export const sumDecisionTypes = (results: NullableDecision[]) => {
-  const stepAccumulation: ResultDecisionCounts = {
-    [NO_DECISION]: 0,
-    [AssessmentDecision.Hold]: 0,
-    [AssessmentDecision.Successful]: 0,
-    [AssessmentDecision.Unsuccessful]: 0,
-  };
-
-  return results.reduce((accumulator: ResultDecisionCounts, result) => {
-    return {
-      ...accumulator,
-      [result]: accumulator[result] + 1,
-    };
-  }, stepAccumulation);
-};
-
 export const decisionOrder: NullableDecision[] = [
   NO_DECISION,
   AssessmentDecision.Successful,
@@ -198,196 +182,6 @@ export const sortResultsAndAddOrdinal = (
       compareFirstNames(resultA, resultB)
     );
   });
-};
-
-const getResultsDecision = (
-  step: AssessmentStep,
-  results?: AssessmentResult[],
-): NullableDecision => {
-  if (!results) return NO_DECISION;
-  let hasFailure: boolean = false;
-  let hasOnHold: boolean = false;
-  let hasToAssess: boolean = false;
-
-  const stepResults = results.filter((result) => {
-    return result.assessmentStep?.id === step.id;
-  });
-
-  if (stepResults.length === 0) {
-    hasToAssess = true;
-  }
-
-  const requiredSkillAssessments = step.poolSkills?.filter(
-    (poolSkill) => poolSkill?.type === PoolSkillType.Essential,
-  );
-
-  requiredSkillAssessments?.forEach((skillAssessment) => {
-    const assessmentResults = stepResults.filter((result) => {
-      return result.poolSkill?.id === skillAssessment?.id;
-    });
-
-    if (assessmentResults.length === 0) {
-      hasToAssess = true;
-      return;
-    }
-
-    assessmentResults.forEach((assessmentResult) => {
-      switch (assessmentResult.assessmentDecision) {
-        case null:
-          hasToAssess = true;
-          break;
-        case AssessmentDecision.Hold:
-          hasOnHold = true;
-          break;
-        case AssessmentDecision.Unsuccessful:
-          hasFailure = true;
-          break;
-        default:
-      }
-    });
-  });
-
-  // Check for Education requirement if this is an ApplicationScreening step
-  if (step.type === AssessmentStepType.ApplicationScreening) {
-    const educationResults = stepResults.filter(
-      (result) =>
-        result.assessmentResultType === AssessmentResultType.Education,
-    );
-    if (educationResults.length === 0) {
-      hasToAssess = true;
-    }
-    educationResults.forEach((result) => {
-      // Any "to assess" should be marked
-      if (result.assessmentDecision === null) {
-        hasToAssess = true;
-      }
-      switch (result.assessmentDecision) {
-        case null:
-          hasToAssess = true;
-          break;
-        case AssessmentDecision.Hold:
-          hasOnHold = true;
-          break;
-        case AssessmentDecision.Unsuccessful:
-          hasFailure = true;
-          break;
-        default:
-      }
-    });
-  }
-
-  if (hasFailure) {
-    return AssessmentDecision.Unsuccessful;
-  }
-  if (hasToAssess) {
-    return NO_DECISION;
-  }
-  if (hasOnHold) {
-    return AssessmentDecision.Hold;
-  }
-
-  return AssessmentDecision.Successful;
-};
-
-type PoolCandidateId = string;
-type AssessmentStepId = string;
-
-export const determineCandidateStatusPerStep = (
-  poolCandidates: PoolCandidate[],
-  steps: AssessmentStep[],
-): Map<PoolCandidateId, Map<AssessmentStepId, NullableDecision>> => {
-  return poolCandidates.reduce((candidateToResults, candidate) => {
-    const assessmentToResult = steps.reduce((map, step) => {
-      return map.set(
-        step.id,
-        getResultsDecision(step, unpackMaybes(candidate.assessmentResults)),
-      );
-    }, new Map<AssessmentStepId, NullableDecision>());
-    candidateToResults.set(candidate.id, assessmentToResult);
-    return candidateToResults;
-  }, new Map<PoolCandidateId, Map<AssessmentStepId, NullableDecision>>());
-};
-
-/**
- * Returns the "current step" a candidate should appear at in the assessment tracker.
- * A candidate should appear in all columns less than or equal to its current step.
- * A value of null means that a candidate has passed all steps.
- * @param assessmentToResult
- * @param assessmentOrdering
- * @returns
- */
-const determineCurrentStep = (
-  assessmentToResult: Map<AssessmentStepId, NullableDecision>,
-  assessmentOrdering: string[],
-): number | null => {
-  for (let index = 0; index < assessmentOrdering.length; index += 1) {
-    const assessmentStepId = assessmentOrdering[index];
-    const result = assessmentToResult.get(assessmentStepId);
-    if (result === AssessmentDecision.Unsuccessful || result === NO_DECISION) {
-      return index;
-    }
-  }
-  return null;
-};
-
-export const determineCurrentStepPerCandidate = (
-  candidateToResults: Map<
-    PoolCandidateId,
-    Map<AssessmentStepId, NullableDecision>
-  >,
-  assessmentSteps: AssessmentStep[],
-): Map<PoolCandidateId, number | null> => {
-  const orderedStepIds = assessmentSteps
-    .sort((stepA, stepB) => {
-      return (stepA.sortOrder ?? Number.MAX_SAFE_INTEGER) >
-        (stepB.sortOrder ?? Number.MAX_SAFE_INTEGER)
-        ? 1
-        : -1;
-    })
-    .map((step) => step.id);
-
-  const candidateToCurrentStep = new Map<PoolCandidateId, number | null>();
-  candidateToResults.forEach((assessmentToResult, candidateId) => {
-    candidateToCurrentStep.set(
-      candidateId,
-      determineCurrentStep(assessmentToResult, orderedStepIds),
-    );
-  });
-  return candidateToCurrentStep;
-};
-
-export const getDecisionCountForEachStep = (
-  assessmentSteps: AssessmentStep[],
-  candidateToResults: Map<
-    PoolCandidateId,
-    Map<AssessmentStepId, NullableDecision>
-  >,
-  candidateToCurrentStep: Map<PoolCandidateId, number | null>,
-): Map<AssessmentStepId, ResultDecisionCounts> => {
-  const orderedSteps = sortBy(assessmentSteps, (step) => step.sortOrder);
-  const decisionCountMap = new Map<AssessmentStepId, ResultDecisionCounts>();
-  for (let index = 0; index < orderedSteps.length; index += 1) {
-    const stepId = orderedSteps[index].id;
-    const poolCandidateIds = Array.from(candidateToCurrentStep.keys());
-    const decisionsForCurrentStep = poolCandidateIds
-      .filter((candidateId) => {
-        // A candidate's result should be counted for its current step and any previous steps
-        // A null step indicates the candidate has successfully passed all assessment steps and should be counted in all of them
-        const candidateStep = candidateToCurrentStep.get(candidateId) ?? null;
-        return candidateStep === null || candidateStep >= index;
-      })
-      // For all the filtered-in candidates, get their result for this step and put them together in an array.
-      .reduce(
-        (decisions: Array<NullableDecision | undefined>, candidateId) => [
-          ...decisions,
-          candidateToResults.get(candidateId)?.get(stepId),
-        ],
-        [],
-      )
-      .filter(notEmpty);
-    decisionCountMap.set(stepId, sumDecisionTypes(decisionsForCurrentStep));
-  }
-  return decisionCountMap;
 };
 
 type StepWithGroupedCandidates = {
